requirements.txt
streamlit
pandas
psycopg2-binary
requests
beautifulsoup4
lxml
plotly
pycountry